+++ packages/FormElement/src/components/Description/Description.js
@@ -1,6 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import * as sc from "./Description.styles";
+import { FormElementContext } from "../../FormElement";
 
 const propTypes = {
   ariaDescriptionId: PropTypes.string,
@@ -12,10 +13,13 @@ const defaultProps = {
 };
 
 function Description(props) {
+  const { formElementA11yProps } = React.useContext(FormElementContext);
   const { children, ariaDescriptionId, ...moreProps } = props;
 
+  const a11yProps = formElementA11yProps ? { id: formElementA11yProps.descriptionA11yProps.id } : {};
+
   return (
-    <sc.Description id={ariaDescriptionId} data-pka-anchor="form-element.description" {...moreProps}>
+    <sc.Description data-pka-anchor="form-element.description" {...moreProps} {...a11yProps}>
       {children}
     </sc.Description>
   );