+++ packages/FormElement/src/index.js
@@ -1,6 +1,6 @@
 export { default as Content } from "./components/Content";
 export { default as Description } from "./components/Description";
-export { default as Error } from "./components/ErrorMessage";
+export { default as Error } from "./components/Error";
 export { default as Help } from "./components/Help";
 export { default as Instructions } from "./components/Instructions";
 export { default as Label } from "./components/Label";