+++ packages/FormElement/stories/examples/ListBoxExample.js
@@ -5,7 +5,6 @@ import FormElement, { useFormElement, Content, Label } from "../../src";
 export default function ListBoxExample() {
   const { listBoxA11yProps, formElementA11yProps } = useFormElement();
   const optionsArray = ["Black Panther", "Wonder Woman", "Spiderman", "The Incredibles", "Thor", <span>test</span>];
-
   const listboxOptions = optionsArray.map(hero => <ListBox.Option key={hero}>{hero}</ListBox.Option>);
 
   return (