+++ packages/FormElement/stories/examples/NativeInputExample.js
@@ -7,12 +7,8 @@ export default function NativeInputExample() {
   const isDisabled = false;
   const isReadOnly = false;
   return (
-    <FormElement
-      isDisabled={isDisabled}
-      hasRequiredLabel={hasRequiredLabel}
-      formElementA11yProps={formElementA11yProps}
-    >
-      <Label>Form Label</Label>
+    <FormElement isDisabled={isDisabled} formElementA11yProps={formElementA11yProps}>
+      <Label hasRequiredLabel={hasRequiredLabel}>Form Label</Label>
       <Content>
         <input aria-required={hasRequiredLabel} disabled={isDisabled} readOnly={isReadOnly} {...inputA11yProps} />
       </Content>