+++ packages/DatePicker/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+2020-10-29
+
+### Removed
+
+- Id prop from root component [@tristanjasper](https://github.com/tristanjasper).
+
 ## [0.5.51] - 2020-09-02
 
 ### Added