+++ packages/FormElement/CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+2020-10-29
+
+### Changed
+
+- Refactor FormElement api [@tristanjasper](https://github.com/tristanjasper).
+
 ## [0.3.0] - 2020-01-17
 
 ### Added